Add exception for loading the database and input
Added exception to handle when the database is not present and a
FileNotFoundError exception is raised. In this case, the program throws
a FileNotFoundError, but with a more helpful message.

Also added exceptions in the input to handle the user not typing ints
where it should have ints. The solution chosen is not raising an excep-
tion, but overwriting the user input with standard values.