Fix bug "categories list growing indefinitely"
The categories list, copied from the categories.pickle database file at
every run, was presenting a behaviour of growing every time a search was
performed. This behaviour was due to the "All categories" category being
inserted repeatedly every time the function get_search_pages was called,
and it didn't reset before doing another search.

This behaviour was corrected by hardcoding in ML_extract_categories.py
that the category "All categories" of ID 0.0 will always be inserted in
position 0 of the categories list, which is stored in the categories.pic
kle file, the database. ML_extract_categories.py was run once as well,
to generated an updated categories database with these specifications.

The line that inserted the formerly mentioned category was deleted also.