Remove tests for printing routines and reorder tests
Removed tests for functions that print due to difficulties implementing
sys.stdout redirection. This was done also because printing routines may
change greatly without much limitation, as some changes could improve
user interface, so tests would need to be rewritten with every change.

The reordering was done to enable one test to use another one that comes
before if necessary, without losing the ability to locate the root of
the problem, as tests that do not rely on others and that will be used
after, will have been confirmed to be working before.