Write docstrings and signatures of unittests
Wrote the signature and documentation of the tests for get_link, as well
as for the whole testCase. Wrote the documentation for TestGetCat and
its tests. Changed name of CATS to backup_CATS, because after importing
ML_scraper.py, the CATS database is accessible as ML_scraper.CATS. A
backup of the initial state of the database allows for testing of its
integrity, and restoring its contents if something changes.

In this commit, a documentation style very similar to NumPy Style was a-
dopted, and will be maintained for the documentation of the entire pro-
ject. In tests, specifically, a summary of what tests are performed for
each testCase class needs to be immediately below the title, and other
sections are to be used if the information provided was not sufficient,
or if there is too much information in the summary.

In the docstrings of the testCase class for the categories database
(CATS), some insignificant changes were made, and some formatting chan-
ges were made just to comply to the new documentation style.