Update example product and write all remaining docstrings
Some minor changes were also made in unittest.py, such as bringing the
product example to outside an specific testCase class (because other
testCases need to be able to read its contents).

The example itself was changed for one that may continue valid for much
longer that the removed example. The listing now is of an iPhone 11, and
iPhones usually continue to be sold new for years after their launch da-
te. This makes possible to test is_reputable more thoroughly.

All documentation for testCases already declared was written in their
first version, and nothing inside unittests.py lacks documentation now,
but the module itself. Nothing was implemented in this commit.