Write tests for get_link, get_title, get_picture
The test that checks that get_link strips irrelevant information from
the returned link is now implemented using 're.search'. A new test for
get_link has been declared but not implemented, which tests if the
string returned is a hyperlink.

In TestGetTitle, all tests are now implemented and documented.

In TestGetPicture, the test for the return type is now implemented, and
the remaining tests are declared but not implemented.