Initial commit of branch: duplicate unittests.py
The branch new-spec-on-tests will be used to introduce the changes in
the specification of the behaviour of the functions that the unittests
module bring (this is an attempt at test-driven development). Most are
not huge changes.

This commit simply duplicates unittests.py, removing from it the tests
that require doing a search in MercadoLivre, which for some reason that
is yet to be identified, inside the unittests module does not seem to
complete for hours. Besides those tests, the others work fine and will
provide adequate testing on changes that are introduced to the code from
now on.